Move various zoom functions to ui
